requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.1.0
google-generativeai==0.3.1
numpy==1.26.0
markdown==3.4.3
supabase==2.3.1
gunicorn==21.2.0 